Move MenuItem to its own file.
